// src/interfaces/formualrioRegistroType.tsx
export type formularioRegistroType = {
    nombre: string;
    apellidos: string;
    correo: string;
    telefono: string;
    password: string;
    confirmPassword: string;
}

export type postNewUserType = {
    nombre: string;
    apellidos: string;
    correo: string;
    password: string;
    telefono: string;
}